refactor(copilot): split route handler into crates/goose/src/copilot/ library modules
