+++ crates/goose-server/src/openapi.rs
@@ -481,6 +481,8 @@ derive_utoipa!(IconTheme as IconThemeSchema);
         super::routes::tunnel::get_tunnel_status,
         super::routes::copilot::review,
         super::routes::copilot::setup,
+        super::routes::copilot::disconnect,
+        super::routes::copilot::get_status,
         super::routes::copilot::comment,
         super::routes::copilot::get_prefs,
         super::routes::copilot::put_prefs,
@@ -665,12 +667,14 @@ derive_utoipa!(IconTheme as IconThemeSchema);
         super::routes::setup::SetupResponse,
         super::tunnel::TunnelInfo,
         super::tunnel::TunnelState,
-        super::routes::copilot::CopilotReviewRequest,
-        super::routes::copilot::CopilotReviewResponse,
-        super::routes::copilot::CopilotSetupResponse,
-        super::routes::copilot::CopilotCommentRequest,
-        super::routes::copilot::CopilotPrefsRequest,
-        super::routes::copilot::CopilotPrefsResponse,
+        goose::copilot::CopilotReviewRequest,
+        goose::copilot::CopilotReviewResponse,
+        goose::copilot::CopilotSetupResponse,
+        goose::copilot::CopilotStatusResponse,
+        goose::copilot::CopilotDisconnectResponse,
+        goose::copilot::CopilotCommentRequest,
+        goose::copilot::CopilotPrefsRequest,
+        goose::copilot::CopilotPrefsResponse,
         goose::copilot::CopilotPrefs,
         goose::copilot::RoutingPrefs,
         goose::copilot::TriggerPreference,
